增加微信支付 使用第三方来判断是否是手机浏览器 内置PHPExcel 修正表单主题死循环Bug

+++ framework/js/user.js
@@ -122,7 +122,7 @@ function action_wealth(title,wid,uid,unit)
 			}
 			url += "&note="+$.str.encode(note);
 			var val = $("#a_val").val();
-			if(!val || (val && parseInt(val)<=0)){
+			if(!val || (val && parseFloat(val)<=0)){
 				$.dialog.alert(p_lang('请填写数值，数值必须大于0'));
 				return false;
 			}
